修改控件布局，修改下菜单栏的高亮显示，设置界面和控制界面的切换数据加载

+++ ksdControl/AreaViewController.m
@@ -99,7 +99,7 @@ NSIndexPath* areaDidSelectRowAtIndexPath;
     int sectionCount = 0;
     if(tableView == groupTableView)
     {
-        sectionCount =  elementViewController.elementSections.count + 1;
+        sectionCount =  (int)elementViewController.elementSections.count + 1;
     }
     else
     {
